Move stat flow tests to top-level runtime harness and delete wasm duplicates
